pyproject.toml
[tool.poetry]
name = "text-autoencoders"
version = "0.1.0"
description = ""
authors = ["Nicky Pochinkov <work@nicky.pro>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
torch = "==2.5.1"
sonar-space = "==0.3.2"
dacite = ">=1.8.1"
fire = ">=0.7.0"
hydra-core = ">=1.3.2"
importlib-resources = "~=6.4"
numpy = ">=1.21"
polars = ">=1.16.0"
pyarrow = ">=16.1.0"
retrying = ">=1.3.4"
sentence-splitter = ">=1.4"
stopes = { version = ">=2.2.0", extras = ["mono"] }
tensorboard = ">=2.18.0"
torchaudio = ">=2.5.1"
fairseq2n = { version = "==0.3.0rc1", source = "fairseq2" }
fairseq2 = { version = "==0.3.0rc1", extras = ["arrow"], source = "fairseq2" }
accelerate = ">=1.2.0"
bert-score = ">=0.3.13"
editdistance = ">=0.8.1"
jinja2 = ">=3.1.3"
nltk = ">=3.9.1"
rouge-score = ">=0.1.2"
sacrebleu = ">=2.4.3"
scikit-learn = ">=1.5.2"
spacy = ">=3.7.5"
textdescriptives = ">=2.8.2"
tiktoken = ">=0.8.0"
transformers = ">=4.45.0"
fairscale = ">=0.4.13"
numba = ">=0.60.0"
en_core_web_sm = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl" }
sacremoses = ">=0.1.1"
scipy = ">=1.14"
datasets = ">=2.18.0"
wtpsplit = ">=2.1.0"
einops = "^0.8.1"
wandb = "^0.19.8"
welford-torch = "^0.2.5"
seaborn = "^0.13.2"
sentence-transformers = "^3.4.1"
plotly = "^6.0.0"
openai = "^1.66.3"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0.0"
ipykernel = "^6.29.5"

[[tool.poetry.source]]
name = "fairseq2"
url = "https://fair.pkg.atmeta.com/fairseq2/whl/rc/pt2.5.1/cu124"
priority = "explicit"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
